feat: export to onnx format + c++ inference code test * add : script to export to onnx format * add : c++ code to run inference of lowpass filter gru * add : evaluation of model C++ inference output in jupyter notebook * feat : correction on dataset creation -> npy files contain now float32 (cpp's float), previously contained float64 (cpp's double)

+++ create_dataset_v2.py
@@ -249,7 +249,7 @@ if __name__ == "__main__":
     idx = 0
     print("Saving samples...")
     for x, y in zip(inputs, outputs):
-        np.save(os.path.join(xpath, f"input-{idx}.npy"), x)
-        np.save(os.path.join(ypath, f"expected-{idx}.npy"), y)
+        np.save(os.path.join(xpath, f"input-{idx}.npy"), x.astype(np.float32))
+        np.save(os.path.join(ypath, f"expected-{idx}.npy"), y.astype(np.float32))
         idx += 1
     print("Samples saved to {}".format(sample_folder))
